mariadb多住一从功能

+++ plugins/mariadb/index.py
@@ -107,6 +107,14 @@ def getDbPort():
     return tmp.groups()[0].strip()
 
 
+def getDbServerId():
+    file = getConf()
+    content = mw.readFile(file)
+    rep = 'server-id\s*=\s*(.*)'
+    tmp = re.search(rep, content)
+    return tmp.groups()[0].strip()
+
+
 def getSocketFile():
     file = getConf()
     content = mw.readFile(file)
@@ -1706,12 +1714,14 @@ def getMasterStatus(version=''):
         data['mode'] = recognizeDbMode()
         data['status'] = master_status
 
-        db = pMysqlDb()
-        dlist = db.query('show slave status')
+        pdb = pMysqlDb('mysql')
+        dlist = pdb.query('show slave status')
+        if len(dlist) < 1:
+            dlist = pdb.query("show all slaves status")
 
-        # print(dlist[0])
-        if len(dlist) > 0 and (dlist[0]["Slave_IO_Running"] == 'Yes' or dlist[0]["Slave_SQL_Running"] == 'Yes'):
-            data['slave_status'] = True
+        for v in dlist:
+            if v["Slave_IO_Running"] == 'Yes' or v["Slave_SQL_Running"] == 'Yes':
+                data['slave_status'] = True
 
         return mw.returnJson(master_status, '设置成功', data)
     except Exception as e:
@@ -1868,8 +1878,13 @@ def getMasterRepSlaveUserCmd(version):
     # 查找同步点
     # SELECT BINLOG_GTID_POS('master1-bin.000002', 561866201);
 
+    sid = getDbServerId()
+    connection_name = ""
+    if sid != '':
+        connection_name = "'r{}' ".format(sid)
+
     if mode == "gtid":
-        sql = "CHANGE MASTER TO MASTER_HOST='" + ip + "', MASTER_PORT=" + port + ", MASTER_USER='" + \
+        sql = "CHANGE MASTER " + connection_name + "TO MASTER_HOST='" + ip + "', MASTER_PORT=" + port + ", MASTER_USER='" + \
             clist[0]['username'] + "', MASTER_PASSWORD='" + \
             clist[0]['password'] + "', " + "MASTER_LOG_FILE='" + mstatus[0]["File"] + \
             "',MASTER_LOG_POS=" + \
@@ -2151,40 +2166,13 @@ def updateSlaveSSH(version=''):
 
 
 def getSlaveList(version=''):
-
     db = pMysqlDb()
     dlist = db.query('show slave status')
-    ret = []
-    for x in range(0, len(dlist)):
-        tmp = {}
-        tmp['Master_User'] = dlist[x]["Master_User"]
-        tmp['Master_Host'] = dlist[x]["Master_Host"]
-        tmp['Master_Port'] = dlist[x]["Master_Port"]
-        tmp['Master_Log_File'] = dlist[x]["Master_Log_File"]
-        tmp['Slave_IO_Running'] = dlist[x]["Slave_IO_Running"]
-        tmp['Slave_SQL_Running'] = dlist[x]["Slave_SQL_Running"]
-        tmp['Last_Error'] = dlist[x]["Last_Error"]
-        tmp['Last_IO_Error'] = dlist[x]["Last_IO_Error"]
-        tmp['Last_SQL_Error'] = dlist[x]["Last_SQL_Error"]
-        tmp['Slave_SQL_Running_State'] = dlist[x]["Slave_SQL_Running_State"]
+    if len(dlist) == 0:
+        dlist = db.query('show all slaves status')
 
-        tmp['Error'] = ''
-        if tmp['Last_Error'] != '':
-            tmp['Error'] = tmp['Last_Error']
-
-        if tmp['Last_IO_Error'] != '':
-            tmp['Error'] = tmp['Last_IO_Error']
-
-        if tmp['Last_SQL_Error'] != '':
-            tmp['Error'] = tmp['Last_SQL_Error']
-
-        if tmp['Error'] == '':
-            tmp['Error'] = tmp['Slave_SQL_Running_State']
-
-        ret.append(tmp)
     data = {}
-    data['data'] = ret
-
+    data['data'] = dlist
     return mw.getJson(data)
 
 
@@ -2208,33 +2196,82 @@ def initSlaveStatus(version=''):
         return initSlaveStatusSyncUser(version)
 
 
+def parseSlaveSyncCmd(cmd):
+    a = {}
+    vlist = cmd.split(',')
+
+    has_connection_name = vlist[0]
+
+    pattern_c = r"CHANGE MASTER \'(.*)\' TO MASTER_HOST"
+    match_val = re.match(pattern_c, has_connection_name, re.I)
+    if match_val:
+        m_groups = match_val.groups()
+        a['Connection_name'] = m_groups[0]
+
+    for i in vlist:
+        tmp = i.strip()
+        tmp_a = tmp.split(" ")
+        real_tmp = tmp_a[len(tmp_a) - 1]
+        kv = real_tmp.split("=")
+        a[kv[0]] = kv[1].replace("'", '').replace("'", '').replace(";", '')
+    return a
+
+
 def initSlaveStatusSyncUser(version=''):
     conn = pSqliteDb('slave_sync_user')
-    data = conn.field('ip,port,user,pass,mode,cmd').find()
-    if len(data) < 1:
+
+    slave_data = conn.field('id,ip,port,user,pass,mode,cmd').select()
+    if len(slave_data) < 1:
         return mw.returnJson(False, '需要先添加同步用户配置!')
 
-    # print(data)
-    db = pMysqlDb()
-    dlist = db.query('show slave status')
-    if len(dlist) > 0:
-        return mw.returnJson(False, '已经初始化好了zz...')
+    pdb = pMysqlDb()
+    if len(slave_data) == 1:
+        dlist = pdb.query('show slave status')
+        if len(dlist) > 0:
+            return mw.returnJson(False, '已经初始化好了zz...')
 
-    u = data
-
-    mode_name = 'classic'
-    if u['mode'] == '1':
-        mode_name = 'gtid'
+    msg = ''
 
     local_mode = recognizeDbMode()
-    if local_mode != mode_name:
-        return mw.returnJson(False, '同步模式不一致!')
+    for x in range(len(slave_data)):
+        slave_t = slave_data[x]
+        base_t = 'IP:' + slave_t['ip'] + ",PORT:" + \
+            slave_t['port'] + ",USER:" + slave_t['user']
 
-    # print(u['cmd'])
-    t = db.query(u['cmd'])
-    # print(t)
-    db.query("start slave")
-    return mw.returnJson(True, '初始化成功!')
+        mode_name = 'classic'
+        if slave_data[x]['mode'] == '1':
+            mode_name = 'gtid'
+
+        if local_mode != mode_name:
+            msg += base_t + '->同步模式不一致'
+            continue
+
+        cmd_sql = slave_t['cmd']
+        if cmd_sql == '':
+            msg += base_t + '->同步命令不能为空'
+            continue
+
+        try:
+            pinfo = parseSlaveSyncCmd(cmd_sql)
+        except Exception as e:
+            return mw.returnJson(False, base_t + '->CMD同步命令不合规范!')
+
+        pdb.query(cmd_sql)
+        if 'Connection_name' in pinfo:
+            # pdb.query("stop slave '{}'".format(pinfo['Connection_name']))
+            pdb.query("start slave '{}'".format(pinfo['Connection_name']))
+        else:
+            pdb.query("start slave")
+
+        # # if len(t) > 0:
+        # #     msg += "Error:" + t[1]
+        # t = pdb.query("start slave")
+        # if len(t) > 0:
+        #     msg += "Error:" + t[1]
+
+    if msg == '':
+        msg = '初始化成功!'
+    return mw.returnJson(True, msg)
 
 
 def initSlaveStatusSSH(version=''):
@@ -2312,61 +2349,39 @@ def setSlaveStatus(version=''):
     if not os.path.exists(mode_file):
         return mw.returnJson(False, '需要先设置同步配置')
 
-    mode = mw.readFile(mode_file)
-    if mode == 'ssh':
-        return setSlaveStatusSSH(version)
-    if mode == 'sync-user':
-        return setSlaveStatusSyncUser(version)
+    pdb = pMysqlDb()
+    dlist = pdb.query('show slave status')
+    if len(dlist) == 0:
+        dlist = pdb.query('show all slaves status')
 
-
-def setSlaveStatusSyncUser(version=''):
-    db = pMysqlDb()
-    dlist = db.query('show slave status')
     if len(dlist) == 0:
         return mw.returnJson(False, '需要手动添加同步账户或者执行初始化!')
 
-    if len(dlist) > 0 and (dlist[0]["Slave_IO_Running"] == 'Yes' or dlist[0]["Slave_SQL_Running"] == 'Yes'):
-        db.query('stop slave')
-    else:
-        ip = dlist[0]['Master_Host']
-        conn = pSqliteDb('slave_sync_user')
-        data = conn.field('ip,port,user,pass,mode,cmd').find()
-        if len(data) == 0:
-            return mw.returnJson(False, '没有数据无法重启!')
-        user = data['user']
-        apass = data['pass']
+    for v in dlist:
+        connection_name = ''
+        cmd = "slave"
+        if 'Connection_name' in v:
+            connection_name = v['Connection_name']
+            cmd = "slave '{}'".format(connection_name)
 
-        db.query("start slave")
-        # db.query("start slave user='{}' password='{}';".format(user, apass))
-
-    return mw.returnJson(True, '设置成功!')
-
-
-def setSlaveStatusSSH(version=''):
-
-    db = pMysqlDb()
-    dlist = db.query('show slave status')
-    if len(dlist) == 0:
-        return mw.returnJson(False, '需要手动添加主服务命令或者执行[初始化]!')
-
-    # print(dlist)
-    if len(dlist) > 0 and (dlist[0]["Slave_IO_Running"] == 'Yes' or dlist[0]["Slave_SQL_Running"] == 'Yes'):
-        db.query('stop slave')
-    else:
-        ip = dlist[0]['Master_Host']
-        conn = pSqliteDb('slave_id_rsa')
-        data = conn.field('ip,ps').where("ip=?", (ip,)).find()
-        if len(data) == 0:
-            return mw.returnJson(False, '没有数据无法重启!')
-        db.query("start slave")
+        if (v["Slave_IO_Running"] == 'Yes' or v["Slave_SQL_Running"] == 'Yes'):
+            pdb.query("stop {}".format(cmd))
+        else:
+            pdb.query("start {}".format(cmd))
 
     return mw.returnJson(True, '设置成功!')
 
 
 def deleteSlave(version=''):
+    args = getArgs()
     db = pMysqlDb()
-    dlist = db.query('stop slave')
-    dlist = db.query('reset slave all')
+    if 'sign' in args:
+        sign = args['sign']
+        db.query("stop slave '{}'".format(sign))
+        db.query("reset slave '{}' all".format(sign))
+    else:
+        db.query('stop slave')
+        db.query('reset slave all')
     return mw.returnJson(True, '删除成功!')
 
 
@@ -2402,12 +2417,51 @@ def dumpMysqlData(version=''):
 
 ############### --- 重要 同步---- ###########
 
-def writeDbSyncStatus(data):
+
+def asyncTmpfile():
     path = '/tmp/db_async_status.txt'
+    return path
+
+
+def writeDbSyncStatus(data):
+    path = asyncTmpfile()
     # status_data['code'] = 1
     # status_data['msg'] = '主服务器备份完成...'
     # status_data['progress'] = 30
     mw.writeFile(path, json.dumps(data))
+
+
+def fullSync(version=''):
+    args = getArgs()
+    data = checkArgs(args, ['db', 'sign', 'begin'])
+    if not data[0]:
+        return data[1]
+
+    status_file = asyncTmpfile()
+    if args['begin'] == '1':
+        cmd = 'cd ' + mw.getRunDir() + ' && python3 ' + \
+            getPluginDir() + \
+            '/index.py do_full_sync {"db":"' + \
+            args['db'] + '","sign":"' + args['sign'] + '"} &'
+        print(cmd)
+        mw.execShell(cmd)
+        return json.dumps({'code': 0, 'msg': '同步数据中!', 'progress': 0})
+
+    if os.path.exists(status_file):
+        c = mw.readFile(status_file)
+        tmp = json.loads(c)
+        if tmp['code'] == 1:
+            sys_dump_sql = "/tmp/dump.sql"
+            if os.path.exists(sys_dump_sql):
+                dump_size = os.path.getsize(sys_dump_sql)
+                tmp['msg'] = tmp['msg'] + ":" + "同步文件:" + mw.toSize(dump_size)
+            c = json.dumps(tmp)
+
+        # if tmp['code'] == 6:
+        #     os.remove(status_file)
+        return c
+
+    return json.dumps({'code': 0, 'msg': '点击开始,开始同步!', 'progress': 0})
 
 
 def doFullSync(version=''):
@@ -2424,16 +2478,23 @@ def doFullSync(version=''):
 
 def doFullSyncUser(version=''):
     args = getArgs()
-    data = checkArgs(args, ['db'])
+    data = checkArgs(args, ['db', 'sign'])
     if not data[0]:
         return data[1]
 
     sync_db = args['db']
+    sync_sign = args['sign']
 
+    # print(sync_sign, sync_db)
     db = pMysqlDb()
 
     conn = pSqliteDb('slave_sync_user')
-    data = conn.field('ip,port,user,pass,mode,cmd').find()
+    if sync_sign != '':
+        data = conn.field('ip,port,user,pass,mode,cmd').where(
+            'ip=?', (sync_sign,)).find()
+    else:
+        data = conn.field('ip,port,user,pass,mode,cmd').find()
+
     user = data['user']
     apass = data['pass']
     port = data['port']
@@ -2442,12 +2503,12 @@ def doFullSyncUser(version=''):
     bak_file = '/tmp/tmp.sql'
 
     writeDbSyncStatus({'code': 0, 'msg': '开始同步...', 'progress': 0})
-
-    writeDbSyncStatus({'code': 1, 'msg': '远程导出数据...', 'progress': 20})
+    writeDbSyncStatus({'code': 1, 'msg': '远程导出数据...', 'progress': 10})
 
     if not os.path.exists(bak_file):
         dump_sql_data = getServerDir() + "/bin/mysqldump --force --opt --default-character-set=utf8 --single-transaction -h" + ip + " -P" + \
             port + " -u" + user + " -p" + apass + " " + sync_db + " > " + bak_file
+        print(dump_sql_data)
         mw.execShell(dump_sql_data)
 
     writeDbSyncStatus({'code': 2, 'msg': '本地导入数据...', 'progress': 40})
@@ -2458,10 +2519,17 @@ def doFullSyncUser(version=''):
             ' ' + sync_db + ' < ' + bak_file
         mw.execShell(my_import_cmd)
 
-    db.query("start slave")
+    pinfo = parseSlaveSyncCmd(data['cmd'])
+    # print(pinfo)
+    if 'Connection_name' in pinfo:
+        db.query("start slave '{}'".format(pinfo['Connection_name']))
+    else:
+        db.query("start slave")
 
     writeDbSyncStatus({'code': 6, 'msg': '从库重启完成...', 'progress': 100})
-    os.system("rm -rf " + bak_file)
+
+    if os.path.exists(bak_file):
+        os.system("rm -rf " + bak_file)
     return True
 
 
@@ -2583,38 +2651,6 @@ def doFullSyncSSH(version=''):
     os.system("rm -rf " + SSH_PRIVATE_KEY)
     os.system("rm -rf /tmp/dump.sql")
     return True
-
-
-def fullSync(version=''):
-    args = getArgs()
-    data = checkArgs(args, ['db', 'begin'])
-    if not data[0]:
-        return data[1]
-
-    status_file = '/tmp/db_async_status.txt'
-    if args['begin'] == '1':
-        cmd = 'cd ' + mw.getRunDir() + ' && python3 ' + \
-            getPluginDir() + \
-            '/index.py do_full_sync {"db":"' + args['db'] + '"} &'
-        print(cmd)
-        mw.execShell(cmd)
-        return json.dumps({'code': 0, 'msg': '同步数据中!', 'progress': 0})
-
-    if os.path.exists(status_file):
-        c = mw.readFile(status_file)
-        tmp = json.loads(c)
-        if tmp['code'] == 1:
-            sys_dump_sql = "/tmp/dump.sql"
-            if os.path.exists(sys_dump_sql):
-                dump_size = os.path.getsize(sys_dump_sql)
-                tmp['msg'] = tmp['msg'] + ":" + "同步文件:" + mw.toSize(dump_size)
-            c = json.dumps(tmp)
-
-        # if tmp['code'] == 6:
-        #     os.remove(status_file)
-        return c
-
-    return json.dumps({'code': 0, 'msg': '点击开始,开始同步!', 'progress': 0})
 
 
 def installPreInspection(version):
